FTUE edits (#11)
Thank you for creating this repo! While using it for running tests on
the `data-modeling` repository, I noticed a couple things for potential
improvement:

1. In `set_session_context!`, `load_context` is called with a keyword
argument which is unexpected as it's defined to take a positional
argument.

2. The error handling in `prepare_package` could be enhanced. While
testing with an invalid `rel-package.json`, the current error
suppression made troubleshooting challenging. I've added a `rethrow`
statement though maybe consider adding an `@error` log instead to help
the user with debugging.

3. There's a small discrepancy in a log message. It references a
`database` field in `struct Config` which doesn't seem to exist.

4. I noticed a minor typo in one of the docstrings.

I hope this PR is helpful. Feel free to take all, some, or none of the
changes.