Adding support to filter tests based on diffs (#12)
Adding some crude change detection mechanism to avoid running all tests
of a package on a PR.

This is useful for `std`, for example, which has multiple namespaces,
each somewhat isolated. There's an extensive test suite for `::std::pkg`
that currently takes some 45 minutes, but if we change something trivial
in `::std::common`, there's usually no need to run that.

This PR gets the list of files that changed and tries to determine which
tests to run using the following heuristic:
- if the file is a test file, run its suite
- if the file is a model file, run the corresponding suite (i.e. if I
modify `model/std/common.rel` we should run the suite under
`test/std/common`).

Note that:
1. if there are _some_ changes but none are related to Rel or Julia files, then all tests are skipped.
2. if there are changes, then only Rel based tests are filtered (Julia based tests are always run)
3. this may miss some relevant tests, e.g. if `::std::pkg` uses
something from `::std::common` that was now modified and broke it. The
`::std::pkg` tests will then fail in the `main` branch, which is not
ideal, but I think can be a good compromise since `main` is not used
directly (it should go to the registry first). We just have to be
careful to not release versions of a library if the main branch is
broken.