Improved heuristics for selecting tests based on changes (#21)
This PR improves the tests that are selected to be executed based on
package changes. In particular:

* we now account for changes to any `.rel` file in a test suite. This
means that changing a file like `before-suite.rel` will cause the whole
suite to run.
* changes to a test file used to cause the whole suite to re-execute.
Now they cause only that test file to re-execute.
* change detection is now also supported for Julia-based tests. Besides
following the same semantics as rel-based tests (changing a model causes
the whole suite to re-execute, changing a file causes only the file to
re-execute), Julia-based tests additionally support a Julia-specific
pattern: if the a model `foo-bar` changes, we search both for a
`test/foo-bar` directory or for a `test/foo_bar_tests.jl` file. This
follows the pattern used by julia tests in other projects, such as the
graphlib.

Because there are a few breaking changes, this also bumps to version
`0.2.0`.